requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.11.2
pandas>=2.1.0
pydantic>=2.0.0
aiosqlite>=0.19.0
email-validator>=2.0.0.post2
webdriver_manager==4.0.1
python-dotenv>=1.0.0